==> Lipidomics/scripts/01_ArrangeData_COF_visitAB_lipidomics.Rmd <==
---
title: "Lipidomics Overfeeding Analysis"
subtitle: "Load and Arrange Raw Data"
author: "Sofie Olund Villumsen"
date: "2023"
output: github_document
---

# Initial settings
Clear workspace
```{r}
rm(list = ls()) 
```

Load packages
```{r message=FALSE}
library(tidyverse)
library(impute) # Imputing data with KNN
```

Set wd
```{r setup}
path <- 'L:/LovbeskyttetMapper/HCOF Stem Cells/Sofie_Databehandling/github/COFpaper_omics_adiposetissue/Lipidomics'
knitr::opts_knit$set(root.dir = path)
```

Load data
```{r message=FALSE}
clinicaldata_selected <- read_csv("L:/LovbeskyttetMapper/HCOF Stem Cells/Sofie_Databehandling/Overfeeding_study/Transcriptomics/data/02_clinicaldata_selected_imputed.csv")

lipiddata <- read_csv("L:/LovbeskyttetMapper/HCOF Stem Cells/Sofie_Databehandling/Baseline_study/Lipidomics/data/02_lipidomics_data_imputed_log2_zscore.csv")

# Get a list of lipid names
lipidnames <- colnames(lipiddata)[3:ncol(lipiddata)]
```

# Prepare lipidomics df for the various analyses
## Lipidomics main analyses
```{r}
lipiddata_ <- lipiddata %>% 
  filter(Label != "C") %>% 
  mutate(Label = case_when(Label == "B." ~ "B",
                           TRUE ~ Label)) %>% 
  left_join(., clinicaldata_selected, by = c("Sample_ID" = "id", "Label" = "visit")) %>% 
  mutate(sample_id = paste("Sample", .$Sample_ID, .$bwcat, .$Label, sep = "_")) %>% 
  arrange(bwcat, NAFLD, Label, Sample_ID) %>% 
  select(sample_id, all_of(lipidnames))
lipiddata_

lipiddata_ %>% 
  write_csv(paste(path, "data/01_lipidomicsdata_visitAB_imputed.csv", sep="/")) 
```


## Lipidomics subgroup analyses (NAFLD)
```{r}
lipiddata_nafld <- lipiddata %>% 
  filter(Label != "C") %>% 
  mutate(Label = case_when(Label == "B." ~ "B",
                           TRUE ~ Label)) %>% 
  left_join(., clinicaldata_selected, by = c("Sample_ID" = "id", "Label" = "visit")) %>% 
  mutate(sample_id = paste("Sample", .$Sample_ID, .$bwcat, .$Label , .$NAFLD, sep = "_")) %>% 
  arrange(bwcat, NAFLD, Label, Sample_ID) %>% 
  select(sample_id, all_of(lipidnames))
lipiddata_nafld

lipiddata_nafld %>% 
  write_csv(paste(path, "data/01_lipidomicsdata_visitAB_nafld_imputed.csv", sep="/"))  
```
