==> Lipidomics/scripts/03_LipidStructureAnalyses_COF_visitAB_lipidomics.Rmd <==
---
title: "Lipidomics Overfeeding Analysis"
subtitle: "Comparion in the Systematical and Structural Composition of the Lipids"
author: "Sofie Olund Villumsen"
date: "2023"
output: 
  html_document:
    toc: true
    toc_depth: 2
---

# Initial settings
Clear workspace
```{r echo=FALSE}
rm(list = ls()) 
```

Load packages
```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(lipidomeR) 
library(webr)
```

Set wd
```{r setup}
path <- 'L:/LovbeskyttetMapper/HCOF Stem Cells/Sofie_Databehandling/github/COFpaper_omics_adiposetissue/Lipidomics'
knitr::opts_knit$set(root.dir = path)
```

Load data
```{r message=FALSE}
clinicaldata_selected <- read_csv("L:/LovbeskyttetMapper/HCOF Stem Cells/Sofie_Databehandling/Overfeeding_study/Transcriptomics/data/02_clinicaldata_selected_imputed.csv")

lipidomicsdata_visitAB <- read_csv(paste(path, "data/01_lipidomicsdata_visitAB_imputed.csv", sep="/"))

lipidomicsdata_visitAB_nafld <- read_csv(paste(path, "data/01_lipidomicsdata_visitAB_nafld_imputed.csv", sep="/"))

lipid_database <- read_csv("L:/LovbeskyttetMapper/HCOF Stem Cells/Sofie_Databehandling/Baseline_study/Lipidomics/data/01_lipid_db_id.csv") %>% select(everything(), -...1)

lipidomicsdata_visitAB_deltacomparison <- read_csv("L:/LovbeskyttetMapper/HCOF Stem Cells/Sofie_Databehandling/github/COFpaper_omics_adiposetissue/Lipidomics/data/02_A5_deltaAB_LBWvsNBW_lipidomics_imputed.csv")
```

# Organize data
```{r}
lipidomicsdata_visitAB_ <- lipidomicsdata_visitAB %>% 
  separate(sample_id, c("Sample", "id", "bwcat", "visit"), sep="_") %>% 
  mutate(sample_id = paste(Sample, .$id, .$bwcat, .$visit, sep = "_"),
         bwcat_visit = paste(.$bwcat, .$visit, sep = "_")) %>% 
  select(sample_id, bwcat, visit, bwcat_visit, everything(), -c("Sample", "id"))
lipidomicsdata_visitAB_
```

```{r}
# Create a mapping of the lipid names
names_mapping <-
   map_lipid_names( x = unique(colnames(lipidomicsdata_visitAB_)[5:ncol(lipidomicsdata_visitAB_)]) )
names_mapping
```



# Lipid composition
```{r}
names_mapping %>% 
  group_by(Class) %>% 
  count() %>% 
  arrange(-n) 

comp_lipids <- names_mapping %>% 
  mutate(Lipid_Class = case_when(Class %in% c("LPE","FA" ,"GlcCer-d","CE","LPC-O",
                                              "PE-P","PA", "dMePE", "DG", "PC-P", 
                                              "PE-O", "CerP-d", "LPC-P", "LPE-P", 
                                              "LacCer-d", "LdMePE") ~ "Other", 
                                 TRUE~Class)) %>% 
  group_by(Lipid_Class) %>% 
  count() %>% 
  rename(n_lipids = n) %>% 
  arrange(-n_lipids) 
comp_lipids
```

## Donut chart
```{r}
donutchart_lipid <- PieDonut(comp_lipids, aes(Lipid_Class, count=n_lipids),  title="Distribution of lipid classes in blood plasma samples from NBW and LBW individuals at baseline and after HCOF",  pieAlpha = 0.7,r0 = 0.7, r1 = 1.3, pieLabelSize = 4.5,
  donutLabelSize = 7)

donutchart_lipid

ggsave(paste(path, "results/03_donutchart_lipidclassdistribution_visitAB_lipidomics.png", sep="/"), plot = donutchart_lipid, device = "png", width = 10, height = 6)

# TIFF file
tiff(file = paste(path, "results/03_donutchart_lipidclassdistribution_visitAB_lipidomics.tiff", sep="/"), units="in", width=10, height=6, res=600, pointsize = 0.0001)
print(donutchart_lipid)
dev.off()
```


# Chains and double bonds
```{r}
names_mapping_ <- names_mapping %>% 
  rename(Lipids = Name) %>% 
  as_tibble()
```

## LBW_A
```{r}
lbw_A_structural_comp <- lipidomicsdata_visitAB_ %>% 
  filter(bwcat_visit == "LBW_A") %>% 
  pivot_longer(-c("sample_id","bwcat","visit","bwcat_visit"), 
               names_to = "Lipids", values_to = "Abundance") %>% 
  left_join(names_mapping_, by = join_by(Lipids)) 

# Number of carbon chains pr. group
lbw_A_carbonchains <- lbw_A_structural_comp %>% 
  group_by(Class, N.carbons) %>% 
  summarise(LBW_A_mean_abundance = mean(Abundance)) 
lbw_A_carbonchains

# Number of double bonds pr. group
lbw_A_doublebonds <- lbw_A_structural_comp %>% 
  group_by(Class, N.double.bonds) %>% 
  summarise(LBW_A_mean_abundance = mean(Abundance))  
lbw_A_doublebonds
```

## LBW_B
```{r}
lbw_B_structural_comp <- lipidomicsdata_visitAB_ %>% 
  filter(bwcat_visit == "LBW_B") %>% 
  pivot_longer(-c("sample_id","bwcat","visit","bwcat_visit"), 
               names_to = "Lipids", values_to = "Abundance") %>% 
  left_join(names_mapping_, by = join_by(Lipids)) 

# Number of carbon chains pr. group
lbw_B_carbonchains <- lbw_B_structural_comp %>% 
  group_by(Class, N.carbons) %>% 
  summarise(LBW_B_mean_abundance = mean(Abundance)) 
lbw_B_carbonchains

# Number of double bonds pr. group
lbw_B_doublebonds <- lbw_B_structural_comp %>% 
  group_by(Class, N.double.bonds) %>% 
  summarise(LBW_B_mean_abundance = mean(Abundance))  
lbw_B_doublebonds
```

## NBW_A
```{r}
nbw_A_structural_comp <- lipidomicsdata_visitAB_ %>% 
  filter(bwcat_visit == "NBW_A") %>% 
  pivot_longer(-c("sample_id","bwcat","visit","bwcat_visit"), 
               names_to = "Lipids", values_to = "Abundance") %>% 
  left_join(names_mapping_, by = join_by(Lipids)) 

# Number of carbon chains pr. group
nbw_A_carbonchains <- nbw_A_structural_comp %>% 
  group_by(Class, N.carbons) %>% 
  summarise(NBW_A_mean_abundance = mean(Abundance)) 
nbw_A_carbonchains

# Number of double bonds pr. group
nbw_A_doublebonds <- nbw_A_structural_comp %>% 
  group_by(Class, N.double.bonds) %>% 
  summarise(NBW_A_mean_abundance = mean(Abundance))  
nbw_A_doublebonds
```

## NBW_B
```{r}
nbw_B_structural_comp <- lipidomicsdata_visitAB_ %>% 
  filter(bwcat_visit == "NBW_B") %>% 
  pivot_longer(-c("sample_id","bwcat","visit","bwcat_visit"), 
               names_to = "Lipids", values_to = "Abundance") %>% 
  left_join(names_mapping_, by = join_by(Lipids)) 

# Number of carbon chains pr. group
nbw_B_carbonchains <- nbw_B_structural_comp %>% 
  group_by(Class, N.carbons) %>% 
  summarise(NBW_B_mean_abundance = mean(Abundance)) 
nbw_B_carbonchains

# Number of double bonds pr. group
nbw_B_doublebonds <- nbw_B_structural_comp %>% 
  group_by(Class, N.double.bonds) %>% 
  summarise(NBW_B_mean_abundance = mean(Abundance))  
nbw_B_doublebonds
```



# Plot: Carbon chains
```{r}
# Combine all the tables
cc_LBW <- full_join(lbw_A_carbonchains,lbw_B_carbonchains, by = join_by("Class", "N.carbons"))
cc_NBW <- full_join(nbw_A_carbonchains,nbw_B_carbonchains, by = join_by("Class", "N.carbons"))
cc_table <- full_join(cc_LBW, cc_NBW, by = join_by("Class", "N.carbons"))
cc_table

# Plot number of carbon chains pr. group
cc_plot <- ggplot(data=cc_table) +
  geom_line(aes(x=N.carbons, y=LBW_A_mean_abundance, group=1), color="darkgreen", size = 0.5)+
  geom_point(aes(x=N.carbons, y=LBW_A_mean_abundance, group=1), color="darkgreen", size = 0.5) +
  
  geom_line(aes(x=N.carbons, y=LBW_B_mean_abundance, group=1), color="lightgreen", size = 0.5)+
  geom_point(aes(x=N.carbons, y=LBW_B_mean_abundance, group=1), color="lightgreen", size = 0.5) +
  
  geom_line(aes(x=N.carbons, y=NBW_A_mean_abundance, group=1), color="blue", size = 0.5)+
  geom_point(aes(x=N.carbons, y=NBW_A_mean_abundance, group=1), color="blue", size = 0.5) +
  
  geom_line(aes(x=N.carbons, y=NBW_B_mean_abundance, group=1), color="lightblue", size = 0.5)+
  geom_point(aes(x=N.carbons, y=NBW_B_mean_abundance, group=1), color="lightblue", size = 0.5) +

  scale_color_manual(values = c("LBW_A_mean_abundance" = 'darkgreen', "LBW_B_mean_abundance" = 'lightgreen',
                                "NBW_A_mean_abundance" = 'blue', "NBW_B_mean_abundance" = 'lightblue')) +
  
  facet_wrap("Class") + 
  labs(x = "Length of carbon chain", y = "Mean lipid abundance")

cc_plot

ggsave(paste(path, "results/03_carboncaindistribution_allclasses_bwcatvisit_visitAB_lipidomics.png", sep="/"), plot = cc_plot, device = "png", width = 7.5, height = 5)

```

# Plot: Carbon chains - selected lipid groups
```{r}
# Combine all the tables
cc_table_selected <- cc_table %>% 
  filter(Class %in% c("TG", "PC", "LPC", "SM-d"))

# Plot number of carbon chains pr. group
cc_plot_selected <- ggplot(data=cc_table_selected) +
  geom_line(aes(x=N.carbons, y=LBW_A_mean_abundance, group=1), color="darkgreen", size = 0.5)+
  geom_point(aes(x=N.carbons, y=LBW_A_mean_abundance, group=1), color="darkgreen", size = 0.5) +
  
  geom_line(aes(x=N.carbons, y=LBW_B_mean_abundance, group=1), color="lightgreen", size = 0.5)+
  geom_point(aes(x=N.carbons, y=LBW_B_mean_abundance, group=1), color="lightgreen", size = 0.5) +
  
  geom_line(aes(x=N.carbons, y=NBW_A_mean_abundance, group=1), color="blue", size = 0.5)+
  geom_point(aes(x=N.carbons, y=NBW_A_mean_abundance, group=1), color="blue", size = 0.5) +
  
  geom_line(aes(x=N.carbons, y=NBW_B_mean_abundance, group=1), color="lightblue", size = 0.5)+
  geom_point(aes(x=N.carbons, y=NBW_B_mean_abundance, group=1), color="lightblue", size = 0.5) +

  scale_color_manual(values = c("LBW_A_mean_abundance" = 'darkgreen', "LBW_B_mean_abundance" = 'lightgreen',
                                "NBW_A_mean_abundance" = 'blue', "NBW_B_mean_abundance" = 'lightblue')) +
  
  facet_wrap("Class") + 
  labs(x = "Length of carbon chain", y = "Mean lipid abundance")

cc_plot_selected

ggsave(paste(path, "results/03_carboncaindistribution_selectedclasses_bwcatvisit_visitAB_lipidomics.png", sep="/"), plot = cc_plot_selected, device = "png", width = 7.5, height = 5)

```

# Plot: Double bonds
```{r}
# Combine all the tables
db_LBW <- full_join(lbw_A_doublebonds,lbw_B_doublebonds, by = join_by("Class", "N.double.bonds"))
db_NBW <- full_join(nbw_A_doublebonds,nbw_B_doublebonds, by = join_by("Class", "N.double.bonds"))
db_table <- full_join(db_LBW, db_NBW, by = join_by("Class", "N.double.bonds"))
db_table

# Plot number of double bonds pr. group
db_plot <- ggplot(data=db_table) +
  geom_line(aes(x=N.double.bonds, y=LBW_A_mean_abundance, group=1), color="darkgreen", size = 0.5)+
  geom_point(aes(x=N.double.bonds, y=LBW_A_mean_abundance, group=1), color="darkgreen", size = 0.5) +
  
  geom_line(aes(x=N.double.bonds, y=LBW_B_mean_abundance, group=1), color="lightgreen", size = 0.5)+
  geom_point(aes(x=N.double.bonds, y=LBW_B_mean_abundance, group=1), color="lightgreen", size = 0.5) +
  
  geom_line(aes(x=N.double.bonds, y=NBW_A_mean_abundance, group=1), color="blue", size = 0.5)+
  geom_point(aes(x=N.double.bonds, y=NBW_A_mean_abundance, group=1), color="blue", size = 0.5) +
  
  geom_line(aes(x=N.double.bonds, y=NBW_B_mean_abundance, group=1), color="lightblue", size = 0.5)+
  geom_point(aes(x=N.double.bonds, y=NBW_B_mean_abundance, group=1), color="lightblue", size = 0.5) +
  
  facet_wrap("Class") + 
  labs(x = "Number of double-bonds", y = "Mean lipid abundance")

db_plot

ggsave(paste(path, "results/03_doublebonddistribution_allclasses_bwcatvisit_visitAB_lipidomics.png", sep="/"), plot = db_plot, device = "png", width = 7.5, height = 5)
```

# Plot: Double bonds - selected lipid groups
```{r}
# Selected lipids classes
db_table_selected <- db_table %>% 
  filter(Class %in% c("TG", "PC", "LPC", "SM-d"))

# Plot number of double bonds pr. group
db_plot_selected <- ggplot(data=db_table_selected) +
  geom_line(aes(x=N.double.bonds, y=LBW_A_mean_abundance, group=1), color="darkgreen", size = 0.5)+
  geom_point(aes(x=N.double.bonds, y=LBW_A_mean_abundance, group=1), color="darkgreen", size = 0.5) +
  
  geom_line(aes(x=N.double.bonds, y=LBW_B_mean_abundance, group=1), color="lightgreen", size = 0.5)+
  geom_point(aes(x=N.double.bonds, y=LBW_B_mean_abundance, group=1), color="lightgreen", size = 0.5) +
  
  geom_line(aes(x=N.double.bonds, y=NBW_A_mean_abundance, group=1), color="blue", size = 0.5)+
  geom_point(aes(x=N.double.bonds, y=NBW_A_mean_abundance, group=1), color="blue", size = 0.5) +
  
  geom_line(aes(x=N.double.bonds, y=NBW_B_mean_abundance, group=1), color="lightblue", size = 0.5)+
  geom_point(aes(x=N.double.bonds, y=NBW_B_mean_abundance, group=1), color="lightblue", size = 0.5) +
  
  facet_wrap("Class") + 
  labs(x = "Number of double-bonds", y = "Mean lipid abundance")

db_plot_selected

ggsave(paste(path, "results/03_doublebonddistribution_selectedclasses_bwcatvisit_visitAB_lipidomics.png", sep="/"), plot = db_plot_selected, device = "png", width = 7.5, height = 5)
```

# Heatmap: Mean lipid abundance
```{r}
# LBW_A
lbw_A_meanabundance <- lbw_A_structural_comp %>% 
  group_by(Lipids) %>% 
  summarise(mean_abundance = mean(Abundance)) %>% 
  mutate(group = "LBW_A") 

# LBW_B
lbw_B_meanabundance <- lbw_B_structural_comp %>% 
  group_by(Lipids) %>% 
  summarise(mean_abundance = mean(Abundance)) %>% 
  mutate(group = "LBW_B") 

# NBW_A
nbw_A_meanabundance <- nbw_A_structural_comp %>% 
  group_by(Lipids) %>% 
  summarise(mean_abundance = mean(Abundance)) %>% 
  mutate(group = "NBW_A") 

# NBW_B
nbw_B_meanabundance <- nbw_B_structural_comp %>% 
  group_by(Lipids) %>% 
  summarise(mean_abundance = mean(Abundance)) %>% 
  mutate(group = "NBW_B") 

meanabundance_all <- rbind(lbw_A_meanabundance, lbw_B_meanabundance, 
                           nbw_A_meanabundance, nbw_B_meanabundance)

# Heatmap 
heatmap_meanabundance <- ggplot(meanabundance_all, aes(group, Lipids, fill= mean_abundance)) + 
  geom_tile()

heatmap_meanabundance

ggsave(paste(path, "results/03_heatmap_meanabundance_bwcatNBWLBW_visitAB_lipidomics.png", sep="/"), plot = heatmap_meanabundance, device = "png", width = 7.5, height = 10)
```


# Heatmaps with lipid structure information
```{r}
# Df to long format
lipidomicsdata_visitAB_longer <- lipidomicsdata_visitAB_ %>%  
  pivot_longer(!c("sample_id","bwcat","visit","bwcat_visit"), names_to = "Lipid_names", values_to = "Concentrations_zscore")

lipid_range <- c(-3,3)
```

## LBW_A
```{r}
lipidomicsdata_visitAB_longer_LBW_A <- lipidomicsdata_visitAB_longer %>% 
  filter(bwcat_visit == "LBW_A") %>% 
  select(Lipid_names, Concentrations_zscore) 

# Create the lipidomeR heatmap of lipid concentrations.
heatmap_lipidome(
   x = lipidomicsdata_visitAB_longer_LBW_A[, c( "Lipid_names", "Concentrations_zscore" ) ],
   names.mapping = names_mapping,
   class.facet = "wrap",
   x.names = "Lipid_names",
   fill.limits = lipid_range,
   fill.midpoint = 0,
   melt.value.name = "Abundance",
   scales = "free"
)
```


## LBW_B
```{r}
lipidomicsdata_visitAB_longer_LBW_B <- lipidomicsdata_visitAB_longer %>% 
  filter(bwcat_visit == "LBW_B") %>% 
  select(Lipid_names, Concentrations_zscore) 

# Create the lipidomeR heatmap of lipid concentrations.
heatmap_lipidome(
   x = lipidomicsdata_visitAB_longer_LBW_B[, c( "Lipid_names", "Concentrations_zscore" ) ],
   names.mapping = names_mapping,
   class.facet = "wrap",
   x.names = "Lipid_names",
   fill.limits = range(x = lipidomicsdata_visitAB_longer$"Concentrations_zscore"),
   fill.midpoint = 0,
   melt.value.name = "Abundance",
   scales = "free"
)
```


## NBW_A
```{r}
lipidomicsdata_visitAB_longer_NBW_A <- lipidomicsdata_visitAB_longer %>% 
  filter(bwcat_visit == "NBW_A") %>% 
  select(Lipid_names, Concentrations_zscore) 

# Create the lipidomeR heatmap of lipid concentrations.
heatmap_lipidome(
   x = lipidomicsdata_visitAB_longer_NBW_A[, c( "Lipid_names", "Concentrations_zscore" ) ],
   names.mapping = names_mapping,
   class.facet = "wrap",
   x.names = "Lipid_names",
   fill.limits = range(x = lipidomicsdata_visitAB_longer$"Concentrations_zscore"),
   fill.midpoint = 0,
   melt.value.name = "Abundance",
   scales = "free"
)
```


## NBW_B
```{r}
lipidomicsdata_visitAB_longer_NBW_B <- lipidomicsdata_visitAB_longer %>% 
  filter(bwcat_visit == "NBW_B") %>% 
  select(Lipid_names, Concentrations_zscore) 

# Create the lipidomeR heatmap of lipid concentrations.
heatmap_lipidome(
   x = lipidomicsdata_visitAB_longer_NBW_B[, c( "Lipid_names", "Concentrations_zscore" ) ],
   names.mapping = names_mapping,
   class.facet = "wrap",
   x.names = "Lipid_names",
   fill.limits = range(x = lipidomicsdata_visitAB_longer$"Concentrations_zscore"),
   fill.midpoint = 0,
   melt.value.name = "Abundance",
   scales = "free"
)
```




# Limma models

## A3: LBW
```{r}
lipidomicsdata_visitAB_LBW <- lipidomicsdata_visitAB_ %>% 
  separate(sample_id, into = c("sample", "id", "bwcat", "visit"), sep = "_") %>% 
  filter(bwcat == "LBW")

# Compute the regression models.
result.limma_lbw <-
   compute_models_with_limma(
       x = lipidomicsdata_visitAB_LBW,
       dependent.variables = names_mapping$"Name",
       independent.variables = c( "bwcat_visit", "id" )
   )
```

```{r}
# Create the figure of all lipids and factors.

figure_output_LBW <-
   heatmap_lipidome_from_limma(
       x = result.limma_lbw$"model",
       names.mapping = names_mapping,
       axis.x.carbons = FALSE,
       class.facet = "row",
       plot.all = FALSE,
       plot.individual = TRUE,
       print.figure = FALSE,
       scales = "free",
       space = "free",
       omit.class = c("PE-P","LPC-O","PC-O", "PC-P", "PG", "PE-O", "GlcCer-d", "FA", "CE", "Cer-d", "PE", "DG", 
                      "dMePE", "CerP-d", "LPE-P", "LPC-P", "PA", "LacCer-d", "LdMePE"),
       p.val.label.bg.size = 5,
       p.val.label.size = 3.5
   )

figure_output_LBW

# Save the heatmap as a PDF file
tiff(file = paste(path, "results/03_heatmap_LipidomeR_A3_LBW_visitAB_lipidomics.tiff", sep="/"),
    width = 25,
    height = 20,
    units = "cm",
    res = 600)

print( figure_output_LBW[[ "bwcat_visitLBW_B" ]] )
dev.off()
```


## A4: NBW
```{r}
lipidomicsdata_visitAB_NBW <- lipidomicsdata_visitAB_ %>% 
  separate(sample_id, into = c("sample", "id", "bwcat", "visit"), sep = "_") %>% 
  filter(bwcat == "NBW")

# Compute the regression models.
result.limma_NBW <-
   compute_models_with_limma(
       x = lipidomicsdata_visitAB_NBW,
       dependent.variables = names_mapping$"Name",
       independent.variables = c( "bwcat_visit", "id" )
   )
```

```{r}
# Create the figure of all lipids and factors.
figure_output_NBW <-
   heatmap_lipidome_from_limma(
       x = result.limma_NBW$"model",
       names.mapping = names_mapping,
       axis.x.carbons = FALSE,
       class.facet = "row",
       plot.all = FALSE,
       plot.individual = TRUE,
       print.figure = FALSE,
       scales = "free",
       space = "free",
       omit.class = c("PE-P","LPC-O","PC-O", "PC-P", "PG", "PE-O", "GlcCer-d", "FA", "CE", "Cer-d", "PE", "DG", 
                      "dMePE", "CerP-d", "LPE-P", "LPC-P", "PA", "LacCer-d", "LdMePE"),
       p.val.label.bg.size = 5,
       p.val.label.size = 3.5
   )

# Save the heatmap as a PDF file
tiff(file = paste(path, "results/03_heatmap_LipidomeR_A4_NBW_visitAB_lipidomics.tiff", sep="/"),
    width = 25,
    height = 20,
    units = "cm",
    res = 600)

print( figure_output_NBW[[ "bwcat_visitNBW_B" ]] )
dev.off()
```









## A5: Delta
```{r}
lipidomicsdata_visitAB_deltacomparison_filtered <- lipidomicsdata_visitAB_deltacomparison %>%
  filter(str_detect(Lipids, "LPC\\(|PC\\(|TG\\(|SM-d\\(")) %>% 
  filter(!str_detect(Lipids, "LPC-O|PC-O"))

# Create the lipidomeR heatmap of lipid concentrations.
heatmap_lipidome_delta <- heatmap_lipidome(
   x = lipidomicsdata_visitAB_deltacomparison_filtered[, c( "Lipids", "logFC" ) ],
   names.mapping = names_mapping,
   class.facet = "wrap",
   x.names = "Lipids",
   fill.limits = c(-0.9,0.9),
   fill.midpoint = 0,
   melt.value.name = "LogFC",
   scales = "free"
)


# Save the heatmap as a PDF file
png(file = paste(path, "results/03_heatmap_LipidomeR_A5_Delta_visitAB_lipidomics.png", sep="/"),
    width = 15,
    height = 10,
    units = "cm",
    res = 300)

print(heatmap_lipidome_delta)
dev.off()
```
