==> Lipidomics/scripts/04_NetworkAnalyses_COF_visitAB_lipidomics.Rmd <==
---
title: "Lipidomics overfeeding Analysis"
subtitle: "Biological Network Topology Analysis"
author: "Sofie Olund Villumsen"
date: "15/05/2023"
output: 
  html_document:
    toc: true
    toc_depth: 2
---

# Initial settings
Clear workspace
```{r echo=FALSE}
rm(list = ls()) 
```

Load packages
```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(correlation)
library(igraph)
library(leidenAlg)
library(lipidomeR)
library(circlize)
```

Set wd
```{r setup}
path <- 'L:/LovbeskyttetMapper/HCOF Stem Cells/Sofie_Databehandling/github/COFpaper_omics_adiposetissue/Lipidomics'
knitr::opts_knit$set(root.dir = path)
```

Load data
```{r}
clinicaldata_selected <- read_csv("L:/LovbeskyttetMapper/HCOF Stem Cells/Sofie_Databehandling/Overfeeding_study/Transcriptomics/data/02_clinicaldata_selected_imputed.csv", show_col_types = FALSE) %>% mutate(sample_id = paste("Sample", .$id, .$bwcat, .$visit, sep = "_")) %>% 
  arrange(bwcat, NAFLD, visit)

lipidomicsdata_visitAB <- read_csv(paste(path,"data/01_lipidomicsdata_visitAB_imputed.csv",sep="/"), show_col_types = FALSE)

lbw_lipidomicsdata_DE <- read_csv(paste(path,"data/03_A3_LBW_visitBvsA_lipidomics_imputed.csv",sep="/"), show_col_types = FALSE) %>% 
  mutate(Lipids = str_replace_all(Lipids, ",", "_"))

nbw_lipidomicsdata_DE <- read_csv(paste(path,"data/03_A4_NBW_visitBvsA_lipidomics_imputed.csv",sep="/"), show_col_types = FALSE) %>% 
  mutate(Lipids = str_replace_all(Lipids, ",", "_"))

lipid_database <- read_csv("L:/LovbeskyttetMapper/HCOF Stem Cells/Sofie_Databehandling/Baseline_study/Lipidomics/data/01_lipid_db_id.csv", show_col_types = FALSE) %>% select(everything(), -...1)

top_lipids <- read_tsv(paste(path, "data/02_toplipids_A5A3A4A10A11_lipidomics.tsv", sep="/"), show_col_types = FALSE)

# Get a list of lipid names
lip_col_names <- colnames(lipidomicsdata_visitAB)[2:ncol(lipidomicsdata_visitAB)]
```

## Structural composition of each community
```{r}
# Create a mapping of the lipid names using lipidomeR package
names_mapping <-
  map_lipid_names( x = unique(colnames(lipidomicsdata_visitAB)[2:ncol(lipidomicsdata_visitAB)]) ) %>% 
  rename(Biochemicals = Name) %>% 
  mutate(Biochemicals = str_replace(Biochemicals, ",", "_")) %>% 
  as_tibble() 
names_mapping
```

Seed: Set a seed for reproducibility of data. Seed sampled with -> sample(1e6, 1)
```{r}
set.seed(465983)
```


# Data preparation
```{r}
lip_df <- lipidomicsdata_visitAB %>% 
  select(-everything(), all_of(lip_col_names)) %>% 
  t()

colnames(lip_df) <- lipidomicsdata_visitAB$sample_id
rownames(lip_df) <- lip_col_names
lip_df <- as_tibble(lip_df) %>% 
  mutate(lipids = lip_col_names) %>% 
  select(lipids, everything()) %>% 
  rename(Lipid = lipids)

kegg <- left_join(lip_df, lipid_database,"Lipid") %>% 
  select(KEGG, HMDB, all_of(colnames(lip_df))) 

write_csv(kegg, paste(path, "data/04_network_lipid_df_pathwayID.csv", sep="/")) 
```



# All: Network clinical + lipid data: Positive + Negative Correlations with abs(rho) > 0.35
Clinical variables selected based on the significant hits (FDR<0.05) from the eigencorplot correlating the PCs with the clinical variables (from PC1 and PC2 with most explained variation)
Clinical variables:
- Hepatic fat
- F-triglyceride
- Total cholesterol
- LDL cholesterol
- HDL cholesterol
- F-C-peptide
- F-insulin
- HOMA-IR
- Total fat mass
- VAT mass
- Weight
- BMI

## Data preparation 
The initial network analysis will be an association analysis using Spearmans rank correlations. The network has a big chance of producing false positives, thus the Benjamin-Hochberg (BH) for false discovery rate (FDR).
```{r}
# LINK: https://statsandr.com/blog/correlation-coefficient-and-correlation-test-in-r/
# LINK github correlation package: https://easystats.github.io/correlation/

# Set variable for the cut-off value of the multiple hypothesis testing (< .001***, < .01**, < .1*)
alpha_thres_clin2 <- 0.1
rho_thres_clin2 <- 0.35

# Scale specific columns in the data frame
scaled_clin2 <- as.data.frame(scale(clinicaldata_selected[, c("Hepatic fat", "F-triglyceride", "Total cholesterol", "LDL cholesterol", "HDL cholesterol","F-C-peptide", "F-insulin", "HOMA-IR", "Total fat mass", "VAT mass", "Weight","BMI")])) %>% 
  mutate(sample_id = clinicaldata_selected$sample_id)

# Combine clinical data with lipid data
biochem_conc_clin2 <- lipidomicsdata_visitAB %>% 
  select(everything()) %>% 
  left_join(scaled_clin2, join_by(sample_id)) %>% 
  select(sample_id, `Hepatic fat`, `F-triglyceride`, `Total cholesterol`, `LDL cholesterol`, `HDL cholesterol`,
         `F-C-peptide`, `F-insulin`, `HOMA-IR`, `Total fat mass`, `VAT mass`, `Weight`,`BMI`, all_of(lip_col_names))
```

## Correlation matrix
```{r}
# OBS: takes several minutes to run!
# Compute correlation (rho) and p-values (BH) for several pairs of variables at once.
corr_matrix_clin2 <- correlation(biochem_conc_clin2, method = "spearman", p_adjust = "BH") %>%  as_tibble()
```


## Edge table: Positive + Negative Correlations
The initial network analysis will be an association analysis using Spearmans rank correlations. The network has a big chance of producing false positives, thus the Benjamin-Hochberg (BH) for false discovery rate (FDR).
```{r}
corr_matrix_all_posneg_clin2 <- corr_matrix_clin2 %>% 
  rename(Biochemical1 = Parameter1, 
         Biochemical2 = Parameter2, 
         pval_adj_BH = p) %>% 
  filter(pval_adj_BH < alpha_thres_clin2) %>% 
  filter(abs(rho) > rho_thres_clin2) %>%  # Threshold of rho is set to be over 0.35 (asolute value of rho)
  select(Biochemical1, Biochemical2, rho) %>% 
  mutate(Biochemical1 = gsub(",", "_", Biochemical1),
         Biochemical2 = gsub(",", "_", Biochemical2)) %>% 
  as.data.frame()

write_csv(corr_matrix_all_posneg_clin2, paste(path, "data/04_network_lipiddata-allvar_clinicaldata-12var_posneg_edgetable_rho35_lipidomics.csv", sep="/"))

# Graph of associations, weighted
g_posneg_all_clin2 <- graph_from_data_frame(corr_matrix_all_posneg_clin2, directed = FALSE)
g_posneg_all_clin2 <- set_edge_attr(g_posneg_all_clin2, "weight", value= corr_matrix_all_posneg_clin2$rho)
is_weighted(g_posneg_all_clin2)

# Plot graph
plot(g_posneg_all_clin2, vertex.label = V(g_posneg_all_clin2)$name)
```

```{r}
# Negative correlations (n=361)
corr_matrix_all_posneg_clin2 %>% 
  filter(rho<0)

# positive correlations (n=12422)
corr_matrix_all_posneg_clin2 %>% 
  filter(rho>0)
```


## Community detection
Community detection techniques are useful to discover biochemicals with common properties and keep them tightly connected
```{r}
#Seed: Set a seed for reproducibility of data. Seed sampled with -> sample(1e6, 1)
set.seed(758474)

# Leidenalg: Community detection algorithm 
leiden_community_all_clin2 <- leiden.community(g_posneg_all_clin2)

# Community seperation
g_posneg_all_community_clin2 <- as_tibble(leiden_community_all_clin2$membership)  %>% 
  mutate(Biochemicals = leiden_community_all_clin2$names) 

# Modularity of network: in a range from -1 to 1, a value close to 1 indicates strong community structure
modularity(g_posneg_all_clin2, membership(leiden_community_all_clin2))

# Number of communities and how many metabolites each contain
sizes(leiden_community_all_clin2)
```

## Node table
```{r}
# Create table which contain information on the metabolites
node_table_all_posneg_clin2 <- g_posneg_all_community_clin2 %>% 
  rename(Community_nr = value) %>% 
  select(Biochemicals, Community_nr)

# Group by Lipids and concatenate comparisons
pvalue_lipids_all_clin2 <- top_lipids %>%
  mutate(Lipid = str_replace_all(Lipids, ",", "_")) %>% 
  group_by(Lipids) %>%
  summarise(concatenated_comparisons = paste(comparison, collapse = ", ")) 

node_table_all_posneg2_clin2 <- left_join(node_table_all_posneg_clin2, pvalue_lipids_all_clin2, by = c("Biochemicals" = "Lipids"), keep = TRUE) 

node_table_all_posneg3_clin2 <- node_table_all_posneg2_clin2 %>% 
  mutate(var_type = case_when(Biochemicals == c("Hepatic fat", "F-triglyceride", "Total cholesterol", 
                                                "LDL cholesterol", "HDL cholesterol","F-C-peptide",
                                                "F-insulin", "HOMA-IR", "Total fat mass",
                                                "VAT mass", "Weight","BMI") ~ "clinical",
                              Biochemicals %in% lip_col_names ~ "lipid",
                              Biochemicals == "PC-O(40:3)_PC(P-40:2)" ~ "lipid",
                              Biochemicals == "PC-O(34:4)_PC(P-34:3)" ~ "lipid",
                              Biochemicals == "PC-O(30:1)_PC(P-30:0)" ~ "lipid",
                              Biochemicals == "PE-O(40:5)_PE(P-40:4)" ~ "lipid",
                              Biochemicals == "PC-O(38:3)_PC(P-38:2)" ~ "lipid",
                              Biochemicals == "LPC-O(18:1)/0:0)_ LPC(P-18:0/0:0)" ~ "lipid",)) %>% 
  select(Biochemicals, Community_nr, concatenated_comparisons, Lipids, var_type) 

node_table_all_posneg3_clin2
write_csv(node_table_all_posneg3_clin2, paste(path, "data/04_network_lipiddata-allvar_clinicaldata-12var_posneg_nodetable_rho35_lipidomics.csv", sep="/"), na="")
```


### C0
```{r}
c0_clin2 <- node_table_all_posneg3_clin2 %>% 
  filter(Community_nr == 0) %>% 
  as_tibble()

c0_structural_comp_clin2 <- c0_clin2 %>% left_join(names_mapping, by = join_by(Biochemicals))
c0_structural_comp_clin2

# Number of lipids pr. lipid class
c0_structural_comp_clin2 %>% 
  group_by(Class) %>% 
  count() %>% 
  arrange(-n)

# plot with double-vond and carbon chains
c0_db_cc_clin2 <- c0_structural_comp_clin2 %>% 
  select(Biochemicals, Class, N.carbons, N.double.bonds)

c0_lipidstructure <- ggplot(data=c0_db_cc_clin2, aes(x=N.double.bonds, y=N.carbons)) +
  geom_point(color="darkgreen") +
  facet_wrap("Class") + 
  ylim(0,65) + 
  xlim(0,15) +
  labs(x = "Number of double-bonds", y = "Number of carbon chains")

c0_lipidstructure

# OBS: NA refers to the clinical variables
ggsave(paste(path, "results/04_network_c0_lipidstructure_visitAB_lipidomics.png", sep="/"), plot = c0_lipidstructure, device = "png", width = 10, height = 6)
```

### C1
```{r}
c1_clin2 <- node_table_all_posneg3_clin2 %>% 
  filter(Community_nr == 1) %>% 
  as_tibble()

c1_structural_comp_clin2 <- c1_clin2 %>% left_join(names_mapping, by = join_by(Biochemicals))
c1_structural_comp_clin2

# Number of lipids pr. lipid class
c1_structural_comp_clin2 %>% 
  group_by(Class) %>% 
  count() %>% 
  arrange(-n)

# plot with double-vond and carbon chains
c1_db_cc_clin2 <- c1_structural_comp_clin2 %>% 
  select(Biochemicals, Class, N.carbons, N.double.bonds)

c1_lipidstructure <- ggplot(data=c1_db_cc_clin2, aes(x=N.double.bonds, y=N.carbons)) +
  geom_point(color="darkgreen") +
  facet_wrap("Class") + 
  ylim(0,65) + 
  xlim(0,15) +
  labs(x = "Number of double-bonds", y = "Number of carbon chains")

c1_lipidstructure

# OBS: NA refers to the clinical variables
ggsave(paste(path, "results/04_network_c1_lipidstructure_visitAB_lipidomics.png", sep="/"), plot = c1_lipidstructure, device = "png", width = 10, height = 6)
```




### C2
```{r}
c2_clin2 <- node_table_all_posneg3_clin2 %>% 
  filter(Community_nr == 2) %>% 
  as_tibble()

c2_structural_comp_clin2 <- c2_clin2 %>% left_join(names_mapping, by = join_by(Biochemicals))
c2_structural_comp_clin2

# Number of lipids pr. lipid class
c2_structural_comp_clin2 %>% 
  group_by(Class) %>% 
  count() %>% 
  arrange(-n)

# plot with double-vond and carbon chains
c2_db_cc_clin2 <- c2_structural_comp_clin2 %>% 
  select(Biochemicals, Class, N.carbons, N.double.bonds)

c2_lipidstructure <- ggplot(data=c2_db_cc_clin2, aes(x=N.double.bonds, y=N.carbons)) +
  geom_point(color="darkgreen") +
  facet_wrap("Class") + 
  ylim(0,65) + 
  xlim(0,15) +
  labs(x = "Number of double-bonds", y = "Number of carbon chains")
c2_lipidstructure

# OBS: NA refers to the clinical variables
ggsave(paste(path, "results/04_network_c2_lipidstructure_visitAB_lipidomics.png", sep="/"), plot = c2_lipidstructure, device = "png", width = 10, height = 6)
```




### C3
```{r}
c3_clin2 <- node_table_all_posneg3_clin2 %>% 
  filter(Community_nr == 3) %>% 
  as_tibble()

c3_structural_comp_clin2 <- c3_clin2 %>% left_join(names_mapping, by = join_by(Biochemicals))
c3_structural_comp_clin2

# Number of lipids pr. lipid class
c3_structural_comp_clin2 %>% 
  group_by(Class) %>% 
  count() %>% 
  arrange(-n)

# plot with double-vond and carbon chains
c3_db_cc_clin2 <- c3_structural_comp_clin2 %>% 
  select(Biochemicals, Class, N.carbons, N.double.bonds)

c3_lipidstructure <- ggplot(data=c3_db_cc_clin2, aes(x=N.double.bonds, y=N.carbons)) +
  geom_point(color="darkgreen") +
  facet_wrap("Class") + 
  ylim(0,65) + 
  xlim(0,15) +
  labs(x = "Number of double-bonds", y = "Number of carbon chains")
c3_lipidstructure

# OBS: NA refers to the clinical variables
ggsave(paste(path, "results/04_network_c3_lipidstructure_visitAB_lipidomics.png", sep="/"), plot = c3_lipidstructure, device = "png", width = 10, height = 6)
```

### C4
```{r}
c4_clin2 <- node_table_all_posneg3_clin2 %>% 
  filter(Community_nr == 4) %>% 
  as_tibble()

c4_structural_comp_clin2 <- c4_clin2 %>% left_join(names_mapping, by = join_by(Biochemicals))
c4_structural_comp_clin2

# Number of lipids pr. lipid class
c4_structural_comp_clin2 %>% 
  group_by(Class) %>% 
  count() %>% 
  arrange(-n)

# plot with double-vond and carbon chains
c4_db_cc_clin2 <- c4_structural_comp_clin2 %>% 
  select(Biochemicals, Class, N.carbons, N.double.bonds)

c4_lipidstructure <- ggplot(data=c4_db_cc_clin2, aes(x=N.double.bonds, y=N.carbons)) +
  geom_point(color="darkgreen") +
  facet_wrap("Class") + 
  ylim(0,65) + 
  xlim(0,15) +
  labs(x = "Number of double-bonds", y = "Number of carbon chains")
c4_lipidstructure

# OBS: NA refers to the clinical variables
ggsave(paste(path, "results/04_network_c4_lipidstructure_visitAB_lipidomics.png", sep="/"), plot = c4_lipidstructure, device = "png", width = 10, height = 6)
```



# NBW: Network clinical + lipid data: Positive + Negative Correlations with abs(rho) > 0.35
Clinical variables selected based on the significant hits (FDR<0.05) from the eigencorplot correlating the PCs with the clinical variables (from PC1 and PC2 with most explained variation)
Clinical variables:
- Hepatic fat
- F-triglyceride
- Total cholesterol
- LDL cholesterol
- HDL cholesterol
- F-C-peptide
- F-insulin
- HOMA-IR
- Total fat mass
- VAT mass
- Weight
- BMI

## Data preparation 
The initial network analysis will be an association analysis using Spearmans rank correlations. The network has a big chance of producing false positives, thus the Benjamin-Hochberg (BH) for false discovery rate (FDR).
```{r}
# LINK: https://statsandr.com/blog/correlation-coefficient-and-correlation-test-in-r/
# LINK github correlation package: https://easystats.github.io/correlation/



# Scale specific columns in the data frame
scaled_clin2_nbw <- as.data.frame(scale(clinicaldata_selected[, c("Hepatic fat", "F-triglyceride", "Total cholesterol", "LDL cholesterol", "HDL cholesterol","F-C-peptide", "F-insulin", "HOMA-IR", "Total fat mass", "VAT mass", "Weight","BMI")])) %>% 
  mutate(sample_id = clinicaldata_selected$sample_id) %>% 
  filter(str_detect(sample_id, "NBW"))

# Combine clinical data with lipid data
biochem_conc_clin2_nbw <- lipidomicsdata_visitAB %>% 
  filter(str_detect(sample_id, "NBW")) %>% 
  select(everything()) %>% 
  left_join(scaled_clin2_nbw, join_by(sample_id)) %>% 
  select(sample_id, `Hepatic fat`, `F-triglyceride`, `Total cholesterol`, `LDL cholesterol`, `HDL cholesterol`,
         `F-C-peptide`, `F-insulin`, `HOMA-IR`, `Total fat mass`, `VAT mass`, `Weight`,`BMI`, all_of(lip_col_names))
```

## Correlation matrix
```{r}
# OBS: takes several minutes to run!
# Compute correlation (rho) and p-values (BH) for several pairs of variables at once.
corr_matrix_clin2_nbw <- correlation(biochem_conc_clin2_nbw, method = "spearman", p_adjust = "BH") %>%  as_tibble()
```


## Edge table: Positive + Negative Correlations
The initial network analysis will be an association analysis using Spearmans rank correlations. The network has a big chance of producing false positives, thus the Benjamin-Hochberg (BH) for false discovery rate (FDR).
```{r}
# Set variable for the cut-off value of the multiple hypothesis testing (< .001***, < .01**, < .1*)
alpha_thres_clin2 <- 0.1
rho_thres_clin2 <- 0.20

corr_matrix_all_posneg_clin2_nbw <- corr_matrix_clin2_nbw %>% 
  rename(Biochemical1 = Parameter1, 
         Biochemical2 = Parameter2, 
         pval_adj_BH = p) %>% 
  filter(pval_adj_BH < alpha_thres_clin2) %>% 
  filter(abs(rho) > rho_thres_clin2) %>%  # Threshold of rho is set to be over 0.35 (asolute value of rho)
  select(Biochemical1, Biochemical2, rho) %>% 
  mutate(Biochemical1 = gsub(",", "_", Biochemical1),
         Biochemical2 = gsub(",", "_", Biochemical2)) %>% 
  as.data.frame()

write_csv(corr_matrix_all_posneg_clin2_nbw, paste(path, "data/04_network_lipiddata-allvar_clinicaldata-12var_posneg_edgetable_rho35_lipidomics_nbw.csv", sep="/"))

# Graph of associations, weighted
g_posneg_all_clin2_nbw <- graph_from_data_frame(corr_matrix_all_posneg_clin2_nbw, directed = FALSE)
g_posneg_all_clin2_nbw <- set_edge_attr(g_posneg_all_clin2_nbw, "weight", value= corr_matrix_all_posneg_clin2_nbw$rho)
is_weighted(g_posneg_all_clin2_nbw)

# Plot graph
plot(g_posneg_all_clin2_nbw, vertex.label = V(g_posneg_all_clin2_nbw)$name)
```

```{r}
# Negative correlations (n=361)
corr_matrix_all_posneg_clin2_nbw %>% 
  filter(rho<0)

# positive correlations (n=12422)
corr_matrix_all_posneg_clin2_nbw %>% 
  filter(rho>0)
```


## Community detection
Community detection techniques are useful to discover biochemicals with common properties and keep them tightly connected
```{r}
#Seed: Set a seed for reproducibility of data. Seed sampled with -> sample(1e6, 1)
set.seed(758474)

# Leidenalg: Community detection algorithm 
leiden_community_all_clin2_nbw <- leiden.community(g_posneg_all_clin2_nbw)

# Community seperation
g_posneg_all_community_clin2_nbw <- as_tibble(leiden_community_all_clin2_nbw$membership)  %>% 
  mutate(Biochemicals = leiden_community_all_clin2_nbw$names) 

# Modularity of network: in a range from -1 to 1, a value close to 1 indicates strong community structure
modularity(g_posneg_all_clin2_nbw, membership(leiden_community_all_clin2_nbw))

# Number of communities and how many metabolites each contain
sizes(leiden_community_all_clin2_nbw)
```

## Node table
```{r}
# Create table which contain information on the metabolites
node_table_all_posneg_clin2_nbw <- g_posneg_all_community_clin2_nbw %>% 
  rename(Community_nr = value) %>% 
  select(Biochemicals, Community_nr)

# Group by Lipids and concatenate comparisons
pvalue_lipids_all_clin2_nbw <- top_lipids %>%
  filter(str_detect(comparison, "_NBW_")) %>% 
  mutate(Lipid = str_replace_all(Lipids, ",", "_")) %>% 
  group_by(Lipids) %>%
  summarise(concatenated_comparisons = paste(comparison, collapse = ", ")) 

node_table_all_posneg2_clin2_nbw <- left_join(node_table_all_posneg_clin2_nbw, pvalue_lipids_all_clin2_nbw, by = c("Biochemicals" = "Lipids"), keep = TRUE) 

node_table_all_posneg3_clin2_nbw <- node_table_all_posneg2_clin2_nbw %>% 
  left_join(.,nbw_lipidomicsdata_DE, by = c("Biochemicals" = "Lipids"))

node_table_all_posneg4_clin2_nbw <- node_table_all_posneg3_clin2_nbw %>% 
  mutate(var_type = case_when(Biochemicals == c("Hepatic fat", "F-triglyceride", "Total cholesterol", 
                                                "LDL cholesterol", "HDL cholesterol","F-C-peptide",
                                                "F-insulin", "HOMA-IR", "Total fat mass",
                                                "VAT mass", "Weight","BMI") ~ "clinical",
                              Biochemicals %in% lip_col_names ~ "lipid",
                              Biochemicals == "PC-O(40:3)_PC(P-40:2)" ~ "lipid",
                              Biochemicals == "PC-O(34:4)_PC(P-34:3)" ~ "lipid",
                              Biochemicals == "PC-O(30:1)_PC(P-30:0)" ~ "lipid",
                              Biochemicals == "PE-O(40:5)_PE(P-40:4)" ~ "lipid",
                              Biochemicals == "PC-O(38:3)_PC(P-38:2)" ~ "lipid",
                              Biochemicals == "LPC-O(18:1)/0:0)_ LPC(P-18:0/0:0)" ~ "lipid",)) %>% 
  select(Biochemicals, Community_nr, concatenated_comparisons, Lipids, var_type, logFC, logFCregulation, topDE, P.Value, adj.P.Val, HMDB, PubChem, KEGG) 

node_table_all_posneg4_clin2_nbw
write_csv(node_table_all_posneg4_clin2_nbw, paste(path, "data/04_network_lipiddata-allvar_clinicaldata-12var_posneg_nodetable_rho35_lipidomics_nbw.csv", sep="/"), na="")
```


### C0
```{r}
c0_clin2_nbw <- node_table_all_posneg3_clin2_nbw %>% 
  filter(Community_nr == 0) %>% 
  as_tibble()

c0_structural_comp_clin2_nbw <- c0_clin2_nbw %>% left_join(names_mapping, by = join_by(Biochemicals))
c0_structural_comp_clin2_nbw

# Number of lipids pr. lipid class
c0_structural_comp_clin2_nbw %>% 
  group_by(Class) %>% 
  count() %>% 
  arrange(-n)

# plot with double-vond and carbon chains
c0_db_cc_clin2_nbw <- c0_structural_comp_clin2_nbw %>% 
  select(Biochemicals, Class, N.carbons, N.double.bonds)

c0_lipidstructure_nbw <- ggplot(data=c0_db_cc_clin2_nbw, aes(x=N.double.bonds, y=N.carbons)) +
  geom_point(color="darkgreen") +
  facet_wrap("Class") + 
  ylim(0,65) + 
  xlim(0,15) +
  labs(x = "Number of double-bonds", y = "Number of carbon chains")

c0_lipidstructure_nbw

# OBS: NA refers to the clinical variables
ggsave(paste(path, "results/04_network_c0_lipidstructure_visitAB_lipidomics_nbw.png", sep="/"), plot = c0_lipidstructure_nbw, device = "png", width = 10, height = 6)
```

### C1
```{r}
c1_clin2 <- node_table_all_posneg3_clin2_nbw %>% 
  filter(Community_nr == 1) %>% 
  as_tibble()

c1_structural_comp_clin2_nbw <- c1_clin2 %>% left_join(names_mapping, by = join_by(Biochemicals))
c1_structural_comp_clin2_nbw

# Number of lipids pr. lipid class
c1_structural_comp_clin2_nbw %>% 
  group_by(Class) %>% 
  count() %>% 
  arrange(-n)

# plot with double-vond and carbon chains
c1_db_cc_clin2_nbw <- c1_structural_comp_clin2_nbw %>% 
  select(Biochemicals, Class, N.carbons, N.double.bonds)

c1_lipidstructure_nbw <- ggplot(data=c1_db_cc_clin2_nbw, aes(x=N.double.bonds, y=N.carbons)) +
  geom_point(color="darkgreen") +
  facet_wrap("Class") + 
  ylim(0,65) + 
  xlim(0,15) +
  labs(x = "Number of double-bonds", y = "Number of carbon chains")

c1_lipidstructure_nbw

# OBS: NA refers to the clinical variables
ggsave(paste(path, "results/04_network_c1_lipidstructure_visitAB_lipidomics_nbw.png", sep="/"), plot = c1_lipidstructure_nbw, device = "png", width = 10, height = 6)
```




### C2
```{r}
c2_clin2 <- node_table_all_posneg3_clin2_nbw %>% 
  filter(Community_nr == 2) %>% 
  as_tibble()

c2_structural_comp_clin2_nbw <- c2_clin2 %>% left_join(names_mapping, by = join_by(Biochemicals))
c2_structural_comp_clin2_nbw

# Number of lipids pr. lipid class
c2_structural_comp_clin2_nbw %>% 
  group_by(Class) %>% 
  count() %>% 
  arrange(-n)

# plot with double-vond and carbon chains
c2_db_cc_clin2_nbw <- c2_structural_comp_clin2_nbw %>% 
  select(Biochemicals, Class, N.carbons, N.double.bonds)

c2_lipidstructure_nbw <- ggplot(data=c2_db_cc_clin2_nbw, aes(x=N.double.bonds, y=N.carbons)) +
  geom_point(color="darkgreen") +
  facet_wrap("Class") + 
  ylim(0,65) + 
  xlim(0,15) +
  labs(x = "Number of double-bonds", y = "Number of carbon chains")
c2_lipidstructure_nbw

# OBS: NA refers to the clinical variables
ggsave(paste(path, "results/04_network_c2_lipidstructure_visitAB_lipidomics_nbw.png", sep="/"), plot = c2_lipidstructure_nbw, device = "png", width = 10, height = 6)
```




### C3
```{r}
c3_clin2 <- node_table_all_posneg3_clin2_nbw %>% 
  filter(Community_nr == 3) %>% 
  as_tibble()

c3_structural_comp_clin2_nbw <- c3_clin2 %>% left_join(names_mapping, by = join_by(Biochemicals))
c3_structural_comp_clin2_nbw

# Number of lipids pr. lipid class
c3_structural_comp_clin2_nbw %>% 
  group_by(Class) %>% 
  count() %>% 
  arrange(-n)

# plot with double-vond and carbon chains
c3_db_cc_clin2_nbw <- c3_structural_comp_clin2_nbw %>% 
  select(Biochemicals, Class, N.carbons, N.double.bonds)

c3_lipidstructure_nbw <- ggplot(data=c3_db_cc_clin2_nbw, aes(x=N.double.bonds, y=N.carbons)) +
  geom_point(color="darkgreen") +
  facet_wrap("Class") + 
  ylim(0,65) + 
  xlim(0,15) +
  labs(x = "Number of double-bonds", y = "Number of carbon chains")
c3_lipidstructure_nbw

# OBS: NA refers to the clinical variables
ggsave(paste(path, "results/04_network_c3_lipidstructure_visitAB_lipidomics_nbw.png", sep="/"), plot = c3_lipidstructure_nbw, device = "png", width = 10, height = 6)
```

### C4
```{r}
c4_clin2 <- node_table_all_posneg3_clin2_nbw %>% 
  filter(Community_nr == 4) %>% 
  as_tibble()

c4_structural_comp_clin2_nbw <- c4_clin2 %>% left_join(names_mapping, by = join_by(Biochemicals))
c4_structural_comp_clin2_nbw

# Number of lipids pr. lipid class
c4_structural_comp_clin2_nbw %>% 
  group_by(Class) %>% 
  count() %>% 
  arrange(-n)

# plot with double-vond and carbon chains
c4_db_cc_clin2_nbw <- c4_structural_comp_clin2_nbw %>% 
  select(Biochemicals, Class, N.carbons, N.double.bonds)

c4_lipidstructure_nbw <- ggplot(data=c4_db_cc_clin2_nbw, aes(x=N.double.bonds, y=N.carbons)) +
  geom_point(color="darkgreen") +
  facet_wrap("Class") + 
  ylim(0,65) + 
  xlim(0,15) +
  labs(x = "Number of double-bonds", y = "Number of carbon chains")
c4_lipidstructure_nbw

# OBS: NA refers to the clinical variables
ggsave(paste(path, "results/04_network_c4_lipidstructure_visitAB_lipidomics_nbw.png", sep="/"), plot = c4_lipidstructure_nbw, device = "png", width = 10, height = 6)
```


# LBW: Network clinical + lipid data: Positive + Negative Correlations with abs(rho) > 0.35
Clinical variables selected based on the significant hits (FDR<0.05) from the eigencorplot correlating the PCs with the clinical variables (from PC1 and PC2 with most explained variation)
Clinical variables:
- Hepatic fat
- F-triglyceride
- Total cholesterol
- LDL cholesterol
- HDL cholesterol
- F-C-peptide
- F-insulin
- HOMA-IR
- Total fat mass
- VAT mass
- Weight
- BMI

## Data preparation 
The initial network analysis will be an association analysis using Spearmans rank correlations. The network has a big chance of producing false positives, thus the Benjamin-Hochberg (BH) for false discovery rate (FDR).
```{r}
# LINK: https://statsandr.com/blog/correlation-coefficient-and-correlation-test-in-r/
# LINK github correlation package: https://easystats.github.io/correlation/

# Scale specific columns in the data frame
scaled_clin2_lbw <- as.data.frame(scale(clinicaldata_selected[, c("Hepatic fat", "F-triglyceride", "Total cholesterol", "LDL cholesterol", "HDL cholesterol","F-C-peptide", "F-insulin", "HOMA-IR", "Total fat mass", "VAT mass", "Weight","BMI")])) %>% 
  mutate(sample_id = clinicaldata_selected$sample_id) %>% 
  filter(str_detect(sample_id, "LBW"))

# Combine clinical data with lipid data
biochem_conc_clin2_lbw <- lipidomicsdata_visitAB %>% 
  filter(str_detect(sample_id, "LBW")) %>% 
  select(everything()) %>% 
  left_join(scaled_clin2_lbw, join_by(sample_id)) %>% 
  select(sample_id, `Hepatic fat`, `F-triglyceride`, `Total cholesterol`, `LDL cholesterol`, `HDL cholesterol`,
         `F-C-peptide`, `F-insulin`, `HOMA-IR`, `Total fat mass`, `VAT mass`, `Weight`,`BMI`, all_of(lip_col_names))
```

## Correlation matrix
```{r}
# OBS: takes several minutes to run!
# Compute correlation (rho) and p-values (BH) for several pairs of variables at once.
corr_matrix_clin2_lbw <- correlation(biochem_conc_clin2_lbw, method = "spearman", p_adjust = "BH") %>%  as_tibble()
```


## Edge table: Positive + Negative Correlations
The initial network analysis will be an association analysis using Spearmans rank correlations. The network has a big chance of producing false positives, thus the Benjamin-Hochberg (BH) for false discovery rate (FDR).
```{r}
# Set variable for the cut-off value of the multiple hypothesis testing (FDR< .001***, FDR< .01**, FDR< .1*)
alpha_thres_clin2 <- 0.1
rho_thres_clin2 <- 0.20

# Filter 
corr_matrix_all_posneg_clin2_lbw <- corr_matrix_clin2_lbw %>% 
  rename(Biochemical1 = Parameter1, 
         Biochemical2 = Parameter2, 
         pval_adj_BH = p) %>% 
  filter(pval_adj_BH < alpha_thres_clin2) %>% 
  filter(abs(rho) > rho_thres_clin2) %>%  # Threshold of rho is set to be over 0.35 (asolute value of rho)
  select(Biochemical1, Biochemical2, rho) %>% 
  mutate(Biochemical1 = gsub(",", "_", Biochemical1),
         Biochemical2 = gsub(",", "_", Biochemical2)) %>% 
  as.data.frame()

write_csv(corr_matrix_all_posneg_clin2_lbw, paste(path, "data/04_network_lipiddata-allvar_clinicaldata-12var_posneg_edgetable_rho35_lipidomics_lbw.csv", sep="/"))

# Graph of associations, weighted
g_posneg_all_clin2_lbw <- graph_from_data_frame(corr_matrix_all_posneg_clin2_lbw, directed = FALSE)
g_posneg_all_clin2_lbw <- set_edge_attr(g_posneg_all_clin2_lbw, "weight", value= corr_matrix_all_posneg_clin2_lbw$rho)
is_weighted(g_posneg_all_clin2_lbw)

# Plot graph
plot(g_posneg_all_clin2_lbw, vertex.label = V(g_posneg_all_clin2_lbw)$name)
```

```{r}
# Negative correlations (n=361)
corr_matrix_all_posneg_clin2_lbw %>% 
  filter(rho<0)

# positive correlations (n=12422)
corr_matrix_all_posneg_clin2_lbw %>% 
  filter(rho>0)
```


## Node table
```{r}
# Create table which contain information on the metabolites
node_table_all_posneg_clin2_lbw <- g_posneg_all_community_clin2_nbw %>% 
  rename(Community_nr = value) 

# Group by Lipids and concatenate comparisons
pvalue_lipids_all_clin2_lbw <- top_lipids %>%
  filter(!str_detect(comparison, "_NBW_")) %>% 
  mutate(Lipid = str_replace_all(Lipids, ",", "_")) %>% 
  group_by(Lipids) %>%
  summarise(concatenated_comparisons = paste(comparison, collapse = ", ")) 

node_table_all_posneg2_clin2_lbw <- left_join(node_table_all_posneg_clin2_lbw, pvalue_lipids_all_clin2_lbw, by = c("Biochemicals" = "Lipids"), keep = TRUE) 

node_table_all_posneg3_clin2_lbw <- node_table_all_posneg2_clin2_lbw %>% 
  left_join(.,lbw_lipidomicsdata_DE, by = c("Biochemicals" = "Lipids"))

node_table_all_posneg4_clin2_lbw <- node_table_all_posneg3_clin2_lbw %>% 
  mutate(var_type = case_when(Biochemicals == c("Hepatic fat", "F-triglyceride", "Total cholesterol", 
                                                "LDL cholesterol", "HDL cholesterol","F-C-peptide",
                                                "F-insulin", "HOMA-IR", "Total fat mass",
                                                "VAT mass", "Weight","BMI") ~ "clinical",
                              Biochemicals %in% lip_col_names ~ "lipid",
                              Biochemicals == "PC-O(40:3)_PC(P-40:2)" ~ "lipid",
                              Biochemicals == "PC-O(34:4)_PC(P-34:3)" ~ "lipid",
                              Biochemicals == "PC-O(30:1)_PC(P-30:0)" ~ "lipid",
                              Biochemicals == "PE-O(40:5)_PE(P-40:4)" ~ "lipid",
                              Biochemicals == "PC-O(38:3)_PC(P-38:2)" ~ "lipid",
                              Biochemicals == "LPC-O(18:1)/0:0)_ LPC(P-18:0/0:0)" ~ "lipid",)) %>% 
  select(Biochemicals, Community_nr, concatenated_comparisons, Lipids, var_type, logFC, logFCregulation, topDE, P.Value, adj.P.Val, HMDB, PubChem, KEGG) 

node_table_all_posneg4_clin2_lbw
write_csv(node_table_all_posneg4_clin2_lbw, paste(path, "data/04_network_lipiddata-allvar_clinicaldata-12var_posneg_nodetable_rho35_lipidomics_lbw.csv", sep="/"), na="")
```



# Circos plot: NBW
```{r}
selected_lipids_nbw <- top_lipids %>%
  filter(str_detect(comparison, "_delta")) %>% 
  select(Lipids) %>% 
  arrange(Lipids)

selected_var_circos_nbw <- c(selected_lipids_nbw$Lipids, colnames(scaled_clin2_nbw))


# Define the names and colors
safe_colorblind_palette <- c(rep("lightgrey", 16),"#88CCEE", "#CC6677", "#DDCC77", "#117733", "#332288", "#AA4499", 
                             "#44AA99", "#999933", "#882255", "black", "#6699CC", "#661100", "red")

# Create the grid.col vector
grid.col <- setNames(safe_colorblind_palette, selected_var_circos_nbw)


filtered_matrix_circos_nbw <- corr_matrix_all_posneg_clin2_nbw %>%
  filter(Biochemical1 %in% selected_var_circos_nbw & Biochemical2 %in% selected_var_circos_nbw) 

#  mutate(Biochemical1_vartype = case_when(Biochemical1 %in% colnames(scaled_clin2_lbw) ~ "clinical",
 #                                          Biochemical1 %in% selected_lipids_nbw$Lipids ~ "lipid"),
  #       Biochemical2_vartype = case_when(Biochemical2 %in% colnames(scaled_clin2_lbw) ~ "clinical",
   #                                       Biochemical2 %in% selected_lipids_nbw$Lipids ~ "lipid"))

filtered_matrix_circos_nbw
```







```{r}

#vartype = c(structure(filtered_matrix_circos$Biochemical1_vartype, names=filtered_matrix_circos$Biochemical1),
 #         structure(filtered_matrix_circos$Biochemical2_vartype ,names= filtered_matrix_circos$Biochemical2))


# Save the heatmap as a PDF file
png(file = paste(path, "results/04_circosplot_NBWnetwork_A5deltalipids_clinicalvar_lipidomics.png", sep="/"),
    width = 30,
    height = 30,
    units = "cm",
    res = 300)



# Create the Circos plot
par(mar = c(1, 1, 1, 1))

chordDiagram(
  x = filtered_matrix_circos_nbw,
  transparency = 0.5,
  annotationTrack = "grid",
  preAllocateTracks = 0,
  order = selected_var_circos_nbw,
  grid.col = grid.col
)

circos.trackPlotRegion(track.index = 1, panel.fun = function(x, y) {
  xlim = get.cell.meta.data("xlim")
  ylim = get.cell.meta.data("ylim")
  sector.name = get.cell.meta.data("sector.index")
  track.height = 0.25
  circos.text(mean(xlim), ylim[1] + .1, sector.name, facing = "clockwise", niceFacing = TRUE, adj = c(0, 0.5), col = "darkgrey")
}, bg.border = NA)


dev.off()


#circos.trackPlotRegion(track.index = 2, panel.fun = function(x, y) {
#    xlim = get.cell.meta.data("xlim")
#    ylim = get.cell.meta.data("ylim")
#    sector.index = get.cell.meta.data("sector.index")
#    circos.text(mean(xlim), mean(ylim), sector.index, col = "white", cex = 0.6, facing = "inside", niceFacing = TRUE)
#}, bg.border = NA)

#for(b in unique(vartype)) {
#  model = names(vartype[vartype == b])
#  highlight.sector(track.index = 2, 
#    text = b, text.vjust = -1, niceFacing = TRUE)
#}

#circos.clear()
```



# Circos plot: LBW
```{r}
selected_lipids_lbw <- top_lipids %>%
  filter(str_detect(comparison, "_delta")) %>% 
  select(Lipids) %>% 
  arrange(Lipids)

selected_var_circos_lbw <- c(selected_lipids_lbw$Lipids, colnames(scaled_clin2_lbw))

filtered_matrix_circos_lbw <- corr_matrix_all_posneg_clin2_lbw %>%
  filter(Biochemical1 %in% selected_var_circos_lbw & Biochemical2 %in% selected_var_circos_lbw) 

#  mutate(Biochemical1_vartype = case_when(Biochemical1 %in% colnames(scaled_clin2_lbw) ~ "clinical",
 #                                          Biochemical1 %in% selected_lipids_lbw$Lipids ~ "lipid"),
  #       Biochemical2_vartype = case_when(Biochemical2 %in% colnames(scaled_clin2_lbw) ~ "clinical",
   #                                       Biochemical2 %in% selected_lipids_lbw$Lipids ~ "lipid"))

filtered_matrix_circos_lbw
```







```{r}

#vartype = c(structure(filtered_matrix_circos$Biochemical1_vartype, names=filtered_matrix_circos$Biochemical1),
 #         structure(filtered_matrix_circos$Biochemical2_vartype ,names= filtered_matrix_circos$Biochemical2))






# Save the heatmap as a PDF file
png(file = paste(path, "results/04_circosplot_LBWnetwork_A5deltalipids_clinicalvar_lipidomics.png", sep="/"),
    width = 30,
    height = 30,
    units = "cm",
    res = 300)


# Create the Circos plot
par(mar = c(1, 1, 1, 1))

chordDiagram(
  x = filtered_matrix_circos_lbw,
  transparency = 0.5,
  annotationTrack = "grid",
  preAllocateTracks = 0,
  order = selected_var_circos_lbw,
  grid.col = grid.col
)

circos.trackPlotRegion(track.index = 1, panel.fun = function(x, y) {
  xlim = get.cell.meta.data("xlim")
  ylim = get.cell.meta.data("ylim")
  sector.name = get.cell.meta.data("sector.index")
  circos.text(mean(xlim), ylim[1] + .1, sector.name, facing = "clockwise", niceFacing = TRUE, adj = c(0, 0.5), col = "darkgrey")
}, bg.border = NA)


dev.off()


#circos.trackPlotRegion(track.index = 2, panel.fun = function(x, y) {
#    xlim = get.cell.meta.data("xlim")
#    ylim = get.cell.meta.data("ylim")
#    sector.index = get.cell.meta.data("sector.index")
#    circos.text(mean(xlim), mean(ylim), sector.index, col = "white", cex = 0.6, facing = "inside", niceFacing = TRUE)
#}, bg.border = NA)

#for(b in unique(vartype)) {
#  model = names(vartype[vartype == b])
#  highlight.sector(track.index = 2, 
#    text = b, text.vjust = -1, niceFacing = TRUE)
#}

#circos.clear()
```
